Fix  aggregate rel bug
