+++ src/include/function/aggregate/count.h
@@ -8,26 +8,14 @@ namespace function {
 struct CountFunction : public BaseCountFunction {
 
     static void updateAll(uint8_t* state_, common::ValueVector* input, uint64_t multiplicity,
-        storage::MemoryManager* memoryManager) {
-        auto state = reinterpret_cast<CountState*>(state_);
-        if (input->hasNoNullsGuarantee()) {
-            for (auto i = 0u; i < input->state->selVector->selectedSize; ++i) {
-                state->count += multiplicity;
-            }
-        } else {
-            for (auto i = 0u; i < input->state->selVector->selectedSize; ++i) {
-                auto pos = input->state->selVector->selectedPositions[i];
-                if (!input->isNull(pos)) {
-                    state->count += multiplicity;
-                }
-            }
-        }
-    }
+        storage::MemoryManager* memoryManager);
 
     static inline void updatePos(uint8_t* state_, common::ValueVector* input, uint64_t multiplicity,
         uint32_t pos, storage::MemoryManager* memoryManager) {
         reinterpret_cast<CountState*>(state_)->count += multiplicity;
     }
+
+    static void paramRewriteFunc(binder::expression_vector& arguments);
 };
 
 } // namespace function
