+++ src/include/planner/logical_plan/schema.h
@@ -3,6 +3,7 @@
 #include <unordered_map>
 
 #include "binder/expression/expression.h"
+#include "binder/expression/expression_util.h"
 
 namespace kuzu {
 namespace planner {
