+++ src/planner/operator/logical_order_by.cpp
@@ -48,9 +48,9 @@ void LogicalOrderBy::computeFlatSchema() {
 }
 
 std::string LogicalOrderBy::getExpressionsForPrinting() const {
-    auto result = binder::ExpressionUtil::toString(expressionsToOrderBy) + "\n";
+    auto result = binder::ExpressionUtil::toString(expressionsToOrderBy) + " ";
     if (hasLimitNum()) {
-        result += "SKIP " + std::to_string(skipNum) + "\n";
+        result += "SKIP " + std::to_string(skipNum) + " ";
         result += "LIMIT " + std::to_string(limitNum);
     }
     return result;
